[android] Fix mouse wheel reporting code syntax (not tested)

+++ ui/Templates/AndroidStudio/Android Template/app/src/main/java/com/giderosmobile/android/AndroidTemplateActivity.java
@@ -284,7 +284,7 @@ public class AndroidTemplateActivity extends Activity implements OnTouchListener
 	      case MotionEvent.ACTION_SCROLL:
 	  		GiderosApplication app = GiderosApplication.getInstance();
 			if (app != null) {
-				app.onMouseWheel(event.getX(),event.getY(),event.getButtonState(),event.getAxisValue(MotionEvent.AXIS_VSCROLL));
+				app.onMouseWheel((int)event.getX(),(int)event.getY(),event.getButtonState(),event.getAxisValue(MotionEvent.AXIS_VSCROLL));
 	        	return true;
 			}
 	    }
